Simple Payments: deactivate EOY Guided Tour
